Removing hard coded fmod define from .pro Changing float to double for better precision in pitch change calculations.

+++ KaraokeHost/khaudiobackendgstreamer.cpp
@@ -330,7 +330,7 @@ int KhAudioBackendGStreamer::pitchShift()
 
 gfloat getPitchForSemitone(int semitone)
 {
-    float pitch;
+    double pitch;
     if (semitone > 0)
     {
         // shifting up
